refactor: remove rules-processor-factory, use direct instantiation
- Delete src/rules/rules-processor-factory.ts and test file
- Add static methods to RulesProcessor class:
  - isToolSupported() - check tool support
  - getSupportedTools() - get supported tools list
  - create() - create instance (replaces getRulesProcessor)
- Update importer.ts and generator.ts to use RulesProcessor.create()
- Update tests to mock RulesProcessor.create instead of factory
- Follow same pattern as SubagentsProcessor in generate.ts
- Remove unnecessary factory abstraction and caching
